library-system/requirements.txt
cassandra-driver==3.29.2
faker==20.1.0
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
flask==3.0.0
